+++ core/collection.ts
@@ -3,6 +3,7 @@ import { getExtension } from "./utils/path.ts";
 import { labelify } from "./utils/string.ts";
 
 import type {
+  CMSContent,
   Data,
   EntryMetadata,
   EntrySource,
@@ -21,6 +22,11 @@ export interface CollectionOptions {
   views?: string[] | ((data?: Data) => string[] | undefined);
   documentName?: string | ((changes: Data) => string | undefined);
   documentLabel?: Labelizer;
+  transform?: (
+    data: Data,
+    CmsContent: CMSContent,
+    isNew: boolean,
+  ) => void | Promise<void>;
   create?: boolean;
   delete?: boolean;
   edit?: boolean;
@@ -43,6 +49,11 @@ export default class Collection {
   previewUrl?: PreviewUrl;
   views?: string[] | ((data?: Data) => string[] | undefined);
   documentName?: string | ((changes: Data) => string | undefined);
+  transform?: (
+    data: Data,
+    CmsContent: CMSContent,
+    isNew: boolean,
+  ) => void | Promise<void>;
   documentLabel?: Labelizer;
   permissions: Permissions;
 
@@ -56,6 +67,7 @@ export default class Collection {
     this.views = options.views;
     this.documentName = options.documentName;
     this.documentLabel = options.documentLabel;
+    this.transform = options.transform;
     this.permissions = {
       edit: options.edit ?? true,
       create: options.create ?? true,
@@ -96,6 +108,7 @@ export default class Collection {
       label,
       entry: this.#storage.get(name),
       fields: this.#fields,
+      transform: this.transform,
     });
   }
 
@@ -108,6 +121,7 @@ export default class Collection {
       label,
       entry: this.#storage.get(name),
       fields: this.#fields,
+      transform: this.transform,
     });
   }